chore: rename REgionConfig to resolveRegionConfig
