+++ packages/config-resolver/src/regionConfig/index.ts
@@ -1,1 +1,1 @@
-export * from "./RegionConfig";
+export * from "./resolveRegionConfig";
